fix(eslint-compat): add main field to package.json (#256)
This is a suggested fix to improve compat with linting - see reasoning below.

When using `eslint-plugin-import` for eslint (a reasonably popular choice, I
believe), it will give false positives when encountering packages that rely
on `exports` in package.json.

There’s a long-standing issue[^1] to fix this, but as of yet it is not done,
leading to users having to manually ignore those errors.

However, in the scenario where the exported entry point is pointing to the main
file of the package, it’s enough to just list that file under the `main` key to
make the false positive go away.

[^1]: https://github.com/import-js/eslint-plugin-import/issues/1810

NOTE: This is an empty commit because the previous commit did not include the
right semantic tag (6ad47232) (#256)

Co-authored-by: Nihal Gonsalves <nihal@nihalgonsalves.com>